TopFilter and other additions

+++ src/main/scala/nl/svanwouw/trending/types/Slope.scala
@@ -5,7 +5,7 @@ package nl.svanwouw.trending.types
  * Represents the slope of the frequency of occurence of a certain topic in a certain period.
  * @param v The value with run time value type.
  */
-class Slope(val v: Int) extends AnyVal with Serializable {
+class Slope(val v: Double) extends AnyVal with Serializable {
   override def toString = v.toString
 }
